+++ ruby/spec/rspec_helper.rb
@@ -2,9 +2,9 @@
 
 RSpec.configure do |config|
   config.before(:suite) do
-    FileUtils.rm(Dir.glob("#{__dir__}/../.asciidoctor/kroki/diag-*"))
+    FileUtils.rm(Dir.glob("#{__dir__}/../.asciidoctor/kroki/{diag-*,hello-world.*,shared.*}"))
   end
   config.after(:suite) do
-    FileUtils.rm(Dir.glob("#{__dir__}/../.asciidoctor/kroki/diag-*")) unless ENV['DEBUG']
+    FileUtils.rm(Dir.glob("#{__dir__}/../.asciidoctor/kroki/{diag-*,hello-world.*,shared.*}")) unless ENV['DEBUG']
   end
 end